combine box_mode and box_utils, reformat
